Partially implement HOME relearner
https://www.smogon.com/forums/threads/sv-should-allow-low-level-move-learns-from-past-gens.3737265/
https://www.smogon.com/forums/threads/home-move-relearner.3736649/

My understanding of the HOME relearner is that a Pokemon going into SV (or any other HOME-compatible game) can relearn any move it previously knew as long as it can be learned in the target game. So I created a general method using the already implemented movepools of SV, SWSH, and past gens to determine what should be able to pass through. This seems to mostly affect low-level Pokemon that can't learn certain level-up moves, but could pull it from an earlier gen due to it being available as a TM or from Gen 7 level-up. Additionally, any level-up/TM/egg move from Gen 8 should always be legal.

In order to fully complete the implementation of HOME relearner, data for moves from LA will have to be added. (Will continue writing, but as of writing this, I'm tired)